requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.8
pillow>=8.3.0
pandas>=1.3.0
numpy>=1.21.0
spacy>=3.1.0
transformers>=4.9.0
python-multipart>=0.0.5
python-dateutil>=2.8.2
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0